+++ R/gtsummary-package.R
@@ -14,7 +14,7 @@
 # allowing for the use of the dot when piping
 utils::globalVariables(".")
 
-# THIS IS FROM PR #195, won't be required when tidyr >=1.0.0 is deps
-tidyr_has_legacy_nest <- function() {
-  exists("unnest_legacy", asNamespace("tidyr"))
-}
+## usethis namespace: start
+#' @importFrom lifecycle deprecate_soft
+## usethis namespace: end
+NULL