+++ R/tbl_summary.R
@@ -45,8 +45,7 @@
 #' e.g. `sort = list(everything() ~ "frequency")`
 #' @param percent Indicates the type of percentage to return. Must be one of
 #' `"column"`, `"row"`, or `"cell"`. Default is `"column"`.
-#' @param group DEPRECATED. Previously required to work with the 'group'
-#' argument in [add_p]. No longer required.
+#' @param group DEPRECATED. Migrated to [add_p]
 #'
 #' @section select helpers:
 #' \href{http://www.danieldsjoberg.com/gtsummary/articles/tbl_summary.html#select_helpers}{Select helpers}
@@ -121,6 +120,19 @@
 #'     statistic = list(all_continuous() ~ "{mean} ({sd})"),
 #'     digits = list(vars(age) ~ c(0, 1))
 #'   )
+#'
+#' # for convenience, you can also pass named lists to any arguments
+#' # that accept formulas (e.g label, digits, etc.)
+#' trial %>%
+#'    dplyr::select(age, grade, response, trt) %>%
+#'    tbl_summary(
+#'     by = trt,
+#'     label = list(age = "Patient Age"),
+#'     statistic = list(all_continuous() ~ "{mean} ({sd})"),
+#'     digits = list(vars(age) ~ c(0, 1))
+#'   )
+#'
+#'
 #' @section Example Output:
 #' \if{html}{Example 1}
 #'
@@ -129,36 +141,74 @@
 #' \if{html}{Example 2}
 #'
 #' \if{html}{\figure{tbl_summary_ex2.png}{options: width=45\%}}
-#'
+
 tbl_summary <- function(data, by = NULL, label = NULL, statistic = NULL,
-                        digits = NULL, type = NULL, value = NULL, group = NULL,
+                        digits = NULL, type = NULL, value = NULL,
                         missing = c("ifany", "always", "no"),
                         missing_text = "Unknown", sort = NULL,
-                        percent = c("column", "row", "cell")) {
-  # matching arguments
+                        percent = c("column", "row", "cell"),  group = NULL) {
+
+  # converting bare arguments to string ----------------------------------------
+  by <- enquo_to_string(rlang::enquo(by), arg_name = "by")
+
+  # putting arguments in a list to pass to tbl_summary_
+  tbl_summary_args <- as.list(environment())
+
+  # passing arguments to tbl_summary_
+  do.call(tbl_summary_, tbl_summary_args)
+}
+
+#' Standard evaluation version of tbl_summary()
+#'
+#' The `'by ='` argument can be passed as a string, rather than with non-standard
+#' evaluation as in [tbl_summary]. Review the help file for [tbl_summary] fully
+#' documented options and arguments.
+#'
+#' @inheritParams tbl_summary
+#' @export
+tbl_summary_ <- function(data, by = NULL, label = NULL, statistic = NULL,
+                        digits = NULL, type = NULL, value = NULL,
+                        missing = c("ifany", "always", "no"),
+                        missing_text = "Unknown", sort = NULL,
+                        percent = c("column", "row", "cell"), group = NULL) {
+  # matching arguments ---------------------------------------------------------
   missing <- match.arg(missing)
   percent <- match.arg(percent)
 
-  # ungrouping data
+  # ungrouping data ------------------------------------------------------------
   data <- data %>% ungroup()
 
-  # converting bare arguments to string -----------------------------------------------
-  by <- enquo_to_string(rlang::enquo(by), arg_name = "by")
+  # deleting obs with missing by values ----------------------------------------
+  # saving variable labels
+  if (!is.null(by) && sum(is.na(data[[by]])) > 0) {
+    message(glue(
+      "{sum(is.na(data[[by]]))} observations missing `{by}` have been removed. ",
+      "To include these observations, use `forcats::fct_explicit_na()` on `{by}` ",
+      "column before passing to `tbl_summary()`."
+    ))
+    lbls <- purrr::map(data, ~attr(.x, "label"))
+    data <- data[!is.na(data[[by]]), ]
 
-  # deprecation note about group
+    # re-applying labels---I think this will NOT be necessary after dplyr 0.9.0
+    for (i in names(lbls)) {
+      attr(data[[i]], "label") <- lbls[[i]]
+    }
+  }
+
+  # deprecation note about group -----------------------------------------------
   if (!rlang::quo_is_null(rlang::enquo(group))) {
-    stop_defunct(glue(
+    stop(glue(
       "Passing the 'group' argument in 'tbl_summary()' is defunct.\n",
       "Please pass the column in 'add_p()'. For example,\n\n",
       "tbl_summary() %>% add_p(group = varname)"
     ))
   }
 
-  # will return call, and all object passed to in tbl_summary call
+  # will return call, and all object passed to in tbl_summary call -------------
   # the object func_inputs is a list of every object passed to the function
   tbl_summary_inputs <- as.list(environment())
 
-  # removing variables with unsupported variable types from data
+  # removing variables with unsupported variable types from data ---------------
   classes_expected <- c("character", "factor", "numeric", "logical", "integer")
   var_to_remove <-
     map_lgl(data, ~ class(.x) %in% classes_expected %>% any()) %>%
@@ -175,17 +225,17 @@ tbl_summary <- function(data, by = NULL, label = NULL, statistic = NULL,
     ))
   }
 
-  # checking function inputs
+  # checking function inputs ---------------------------------------------------
   tbl_summary_input_checks(
     data, by, label, type, value, statistic,
-    digits, missing, missing_text, group, sort
+    digits, missing, missing_text, sort
   )
 
-  # converting tidyselect formula lists to named lists
+  # converting tidyselect formula lists to named lists -------------------------
   type <- tidyselect_to_list(data, type, input_type = "type")
   value <- tidyselect_to_list(data, value, input_type = "value")
 
-  # creating a table with meta data about each variable
+  # creating a table with meta data about each variable ------------------------
   meta_data <- tibble(
     variable = names(data),
     # assigning class, if entire var is NA, then assigning class NA
@@ -196,16 +246,14 @@ tbl_summary <- function(data, by = NULL, label = NULL, statistic = NULL,
   )
   # excluding by variable
   if (!is.null(by)) meta_data <- meta_data %>% filter(!!parse_expr("variable != by"))
-  # excluding id variable
-  if (!is.null(group)) meta_data <- meta_data %>% filter(!!parse_expr("!variable %in% group"))
 
-  # converting tidyselect formula lists to named lists
+  # converting tidyselect formula lists to named lists -------------------------
   label <- tidyselect_to_list(data, label, .meta_data = meta_data, input_type = "label")
   statistic <- tidyselect_to_list(data, statistic, .meta_data = meta_data, input_type = "statistic")
   digits <- tidyselect_to_list(data, digits, .meta_data = meta_data, input_type = "digits")
   sort <- tidyselect_to_list(data, sort, .meta_data = meta_data)
 
-  # assigning variable characteristics
+  # assigning variable characteristics -----------------------------------------
   meta_data <-
     meta_data %>%
     mutate(
@@ -219,7 +267,7 @@ tbl_summary <- function(data, by = NULL, label = NULL, statistic = NULL,
       sort = assign_sort(.data$variable, .data$summary_type, sort)
     )
 
-  # calculating summary statistics
+  # calculating summary statistics ---------------------------------------------
   table_body <-
     meta_data %>%
     mutate(
@@ -243,12 +291,20 @@ tbl_summary <- function(data, by = NULL, label = NULL, statistic = NULL,
     select(c("variable", "summary_type", "stat_table")) %>%
     unnest(!!sym("stat_table"))
 
-  # table of column headers
+  # table of column headers ----------------------------------------------------
   table_header <-
     tibble(column = names(table_body) %>% setdiff("summary_type")) %>%
-    table_header_fill_missing()
+    table_header_fill_missing() %>%
+    mutate(
+      footnote = map2(
+        .data$column, .data$footnote,
+        function(x, y) {
+          if (x == "label") return(c(y, footnote_stat_label(meta_data))); return(y)
+        }
+      )
+    )
 
-  # returning all results in a list
+  # returning all results in a list --------------------------------------------
   results <- list(
     gt_calls = eval(gt_tbl_summary),
     kable_calls = eval(kable_tbl_summary),
@@ -316,15 +372,6 @@ gt_tbl_summary <- quote(list(
     "columns = gt::vars(label),",
     "rows = row_type != 'label'",
     "))"
-  ),
-
-  # adding footnote listing statistics presented in table
-  footnote_stat_label = glue(
-    "gt::tab_footnote(",
-    "footnote = '{footnote_stat_label(meta_data)}',",
-    "locations = gt::cells_column_labels(",
-    "columns = gt::vars(label))",
-    ")"
   )
 ))