+++ R/utils-tbl_summary.R
@@ -1051,9 +1051,8 @@ calculate_single_stat <- function(x, stat_name) {
 # this should include EVERY input of \code{\link{tbl_summary}} in the same order
 # copy and paste them from \code{\link{tbl_summary}}
 
-tbl_summary_input_checks <- function(data, by, label, type, value,
-                                     statistic, digits, missing, missing_text,
-                                     group, sort) {
+tbl_summary_input_checks <- function(data, by, label, type, value, statistic,
+                                     digits, missing, missing_text, sort) {
   # data -----------------------------------------------------------------------
   # data is a data frame
   if (!is.data.frame(data)) {
@@ -1280,11 +1279,6 @@ tbl_summary_input_checks <- function(data, by, label, type, value,
     stop("Argument 'missing_text' must be a character string of length 1.")
   }
 
-  # group ----------------------------------------------------------------------
-  if (length(group) > 1) {
-    stop("'group' must be `NULL` or length 1.")
-  }
-
   # sort -----------------------------------------------------------------------
   if (!is.null(sort) & is.null(names(sort))) { # checking names for deprecated named list input
 
@@ -1453,7 +1447,7 @@ enquo_to_string <- function(by_enquo, arg_name) {
 
   # is user supplied string, then stopping with error
   if (startsWith(by_quo_text, "\"") && endsWith(by_quo_text, "\"")) {
-    stop_defunct(glue(
+    stop(glue(
       "\nPassing the '{arg_name}' argument as a string is defunct.\n",
       "Please pass the {arg_name} argument without quotes. For example, \n\n",
       "foo({arg_name} = varname)"