+++ inst/doc/tbl_summary.R
@@ -8,6 +8,18 @@ knitr::opts_chunk$set(
 library(gtsummary)
 library(dplyr)
 
+# installing gt 
+if (curl::has_internet()) {
+  # adding tmpdir to libPath
+  temp_path <- file.path(tempdir(), "gt_folder")
+  dir.create(temp_path)
+  lib_path <-.libPaths()
+  .libPaths(c(lib_path, temp_path))
+  
+  # installing gt
+  remotes::install_github("rstudio/gt", lib = temp_path)
+}
+
 ## ---- eval=FALSE---------------------------------------------------------
 #  install.packages("gtsummary")
 #  remotes::install_github("rstudio/gt")
@@ -78,7 +90,7 @@ tbl_summary(trial2) %>% purrr::pluck("gt_calls") %>% head(n = 5)
 
 ## ----as_gt2, eval=FALSE--------------------------------------------------
 #  tbl_summary(trial2, by = trt) %>%
-#    as_gt(exclude = "footnote_stat_label") %>%
+#    as_gt(exclude = "tab_footnote") %>%
 #    gt::tab_spanner(label = "Randomization Group",
 #                    columns = gt::starts_with("stat_"))
 
@@ -86,7 +98,7 @@ tbl_summary(trial2) %>% purrr::pluck("gt_calls") %>% head(n = 5)
 # this code chunk only works if gt is installed
 if (requireNamespace("gt", quietly = TRUE)) {
   tbl_summary(trial2, by = trt) %>%
-    as_gt(exclude = "footnote_stat_label") %>%
+    as_gt(exclude = "tab_footnote") %>%
     gt::tab_spanner(label = "Randomization Group",
                     columns = gt::starts_with("stat_"))
 }
@@ -95,6 +107,20 @@ if (requireNamespace("gt", quietly = TRUE)) {
 #  help("Rprofile")
 #  
 #  usethis::edit_r_profile()
+
+## ---- echo=FALSE---------------------------------------------------------
+data.frame(
+  `Description` = c("Formatting and rounding p-values", 
+                    "Formatting and rounding percentages",
+                    "Print tables with `gt` or `kable`"),
+  `Example` = c("`options(gtsummary.pvalue_fun = function(x) gtsummary::style_pvalue(x, digits = 2))`",
+                '`options(gtsummary.tbl_summary.percent_fun = function(x) sprintf("%.2f", 100 * x))`',
+                '`options(gtsummary.print_engine = "kable")`   `options(gtsummary.print_engine = "gt")`'),
+  `Functions` = c("`add_p()`, `tbl_regression()`, `tbl_uvregression()`",
+                  "`tbl_summary()`",
+                  "All `tbl_*()` functions")
+) %>% 
+  knitr::kable()
 
 ## ----eval=FALSE----------------------------------------------------------
 #  options(gtsummary.tbl_summary.percent_fun = function(x) sigfig(x, digits = 3))